Update .gitignore, refactor logging, and target .NET 8.0
- Updated .gitignore for Visual Studio projects to ignore user-specific files, build results, cache files, and other temporary or auto-generated files.
- Changed `LogToConsole` method in `Form1.cs` from `private` to `internal`.
- Refactored database creation logic in `Form1.cs` to use the new `SqlHelper` class.
- Added `FormOneConsoleLogger` class in `FormOneConsoleLogger.cs` implementing `IQuickLogLogger` for logging to the console.
- Added `IQuickLogLogger` interface in `IQuickLogLogger.cs` defining a method for logging messages.
- Added `SqlHelper` class in `SqlHelper.cs` for database setup and logging.
- Updated `quickLog.csproj` to target .NET 8.0 instead of .NET 6.0.
