+++ quickLog/quickLog/IQuickLogLogger.cs
@@ -1,0 +1,7 @@
+﻿namespace quickLog
+{
+    internal interface IQuickLogLogger
+    {
+        void LogToConsole(string message);
+    }
+}
